pyproject.toml
[project]
name = "dune-rag-bot"
version = "0.1.0"
description = "A Slack chatbot that answers questions using knowledge from Confluence & source code, powered by RAG"
readme = "README.md"
requires-python="==3.13.*"
dependencies = [
    "aiohttp>=3.9.0",
    "atlassian-python-api>=4.0.7",
    "beautifulsoup4>=4.12.0",
    "langchain-chroma>=1.1.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "lxml>=5.0.0",
    "python-dotenv>=1.0.0",
    "slack-bolt>=1.18.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-kotlin>=1.1.0",
    "tree-sitter-ruby>=0.23.1",
    "tree-sitter-swift>=0.0.1",
    "tree-sitter-typescript>=0.23.2",
]

[tool.poe.tasks]
check = "uv run python -m py_compile app/*.py"

[dependency-groups]
dev = [
    "poethepoet>=0.38.0",
    "ty",
]
